requirements.txt
# Core Package
fastapi~=0.115.2
uvicorn~=0.34.3
python-dotenv~=1.1.0
pydantic~=2.11.5
SQLAlchemy==2.0.41
starlette~=0.40.0
python-jose~=3.5.0


# Database
asyncpg==0.30.0
psycopg2-binary==2.9.11
alembic==1.16.1

# Formating files
black~=25.1.0